+++ springy/Utils/SoapClient.php
@@ -6,8 +6,6 @@
  * @copyright 2024 Fernando Val
  * @author    Fernando Val <fernando.val@gmail.com>
  * @license   https://github.com/fernandoval/Springy/blob/master/LICENSE MIT
- *
- * @version 1.0.0
  */
 
 namespace Springy\Utils;
@@ -16,7 +14,6 @@ use Exception;
 use SoapClient as GlobalSoapClient;
 use SoapFault;
 use SoapHeader;
-use Springy\Configuration;
 use Springy\Exceptions\SpringyException;
 use Springy\Kernel;
 
@@ -27,9 +24,21 @@ class SoapClient extends GlobalSoapClient
     public function __construct(?string $wsdl, array $options, array $wsse)
     {
         $options['exceptions'] = $options['exceptions'] ?? true;
-        $this->initProxy($options);
-        $this->initTimeout($options);
-        $this->initWsse($wsse, $options);
+        $options += (
+            config_get('network.soap.proxy.host') ? [
+                'proxy_host' => config_get('network.soap.proxy.host'),
+                'proxy_port' => config_get('network.soap.proxy.port'),
+            ] : []
+        ) + (
+            config_get('network.soap.proxy.user') ? [
+                'proxy_login' => config_get('network.soap.proxy.user'),
+                'proxy_password' => config_get('network.soap.proxy.password'),
+            ] : []
+        ) + (
+            ['connection_timeout' => config_get('network.soap.timeout') ?: 20]
+        ) + (
+            $this->isUsingWsse($wsse) ? ['trace' => true] : []
+        );
 
         $met = ini_get('max_execution_time');
         set_time_limit(0);
@@ -37,7 +46,7 @@ class SoapClient extends GlobalSoapClient
 
         try {
             parent::__construct($wsdl, $options);
-            $this->setWsse($wsse, $options);
+            $this->setWsse($wsse);
             $this->error = null;
         } catch (SoapFault $error) {
             $this->error = new SpringyException(
@@ -110,44 +119,6 @@ class SoapClient extends GlobalSoapClient
     public function getError(): ?SpringyException
     {
         return $this->error;
-    }
-
-    private function initProxy(array &$options): void
-    {
-        $host = Configuration::get('network.soap.proxy.host') ?? Configuration::get('system.proxyhost');
-        $port = Configuration::get('network.soap.proxy.port') ?? Configuration::get('system.proxyport');
-        $user = Configuration::get('network.soap.proxy.user') ?? Configuration::get('system.proxyusername');
-        $pass = Configuration::get('network.soap.proxy.password') ?? Configuration::get('system.proxypassword');
-
-        if (!isset($options['proxy_host']) && $host) {
-            $options['proxy_host'] = $host;
-        }
-        if (!isset($options['proxy_port']) && $port) {
-            $options['proxy_port'] = $port;
-        }
-        if (!isset($options['proxy_login']) && $user) {
-            $options['proxy_login'] = $user;
-        }
-        if (!isset($options['proxy_password']) && $pass) {
-            $options['proxy_password'] = $pass;
-        }
-    }
-
-    private function initTimeout(array &$options): void
-    {
-        if (empty($options['connection_timeout'])) {
-            $options['connection_timeout'] = Configuration::get('network.soap.timeout')
-                ?? Configuration::get('soap.timeout') ?: 20;
-        }
-
-        ini_set('default_socket_timeout', $options['connection_timeout']);
-    }
-
-    private function initWsse(array $wsse, array &$options): void
-    {
-        if ($this->isUsingWsse($wsse)) {
-            $options['trace'] = true;
-        }
     }
 
     private function isUsingWsse(array $wsse): bool